Added data

+++ MTLPERV1/matrixmath.h
@@ -4,8 +4,8 @@
 
 std::vector<double> mat_to_row(std::vector<std::vector<uint8_t>> mat){
     std::vector<double> temp(mat.size()*mat[0].size(),0);
-    for(int i = 0; i < mat.size(); ++i){
-        for(int j = 0; j < mat[i].size(); ++j){
+    for(size_t i = 0; i < mat.size(); ++i){
+        for(size_t j = 0; j < mat[i].size(); ++j){
             temp[(i*mat.size() + j)] = mat[i][j];
         }
     }
@@ -18,9 +18,9 @@ std::vector<double> matrix_mult(std::vector<double> out, std::vector<std::vector
         exit(1);
     }
     std::vector<double> temp(weights.size(), 0);
-    for(int outer = 0; outer < weights.size(); ++outer){
-        for(int inner_mat = 0; inner_mat < weights[outer].size(); ++inner_mat){
-            for(int inner_out = 0; inner_out < out.size(); ++inner_out){
+    for(size_t outer = 0; outer < weights.size(); ++outer){
+        for(size_t inner_mat = 0; inner_mat < weights[outer].size(); ++inner_mat){
+            for(size_t inner_out = 0; inner_out < out.size(); ++inner_out){
                 temp[outer] = weights[outer][inner_mat] * out[inner_out];
             }
         }
